create lib for common modules
