+++ src/cgw-common/src/cgw_device.rs
