+++ src/cgw_connection_processor.rs
@@ -1,7 +1,5 @@
 use crate::{
     cgw_connection_server::{CGWConnectionServer, CGWConnectionServerReqMsg},
-    cgw_device::{CGWDeviceCapabilities, CGWDeviceType},
-    cgw_errors::{Error, Result},
     cgw_nb_api_listener::{
         cgw_construct_cloud_header, cgw_construct_infra_join_msg,
         cgw_construct_infra_realtime_event_message, cgw_construct_infra_request_result_msg,
@@ -12,11 +10,16 @@ use crate::{
         CGWUCentralMessagesQueueItem, CGWUCentralMessagesQueueState, CGW_MESSAGES_QUEUE,
         MESSAGE_TIMEOUT_DURATION,
     },
+    cgw_ucentral_topology_map::CGWUCentralTopologyMap,
+};
+
+use cgw_common::{
+    cgw_errors::{Error, Result},
     cgw_ucentral_parser::{
         cgw_ucentral_event_parse, cgw_ucentral_parse_connect_event, CGWUCentralCommandType,
         CGWUCentralEventType, CGWUCentralReplyType,
     },
-    cgw_ucentral_topology_map::CGWUCentralTopologyMap,
+    cgw_device::{CGWDeviceCapabilities, CGWDeviceType},
 };
 
 use eui48::MacAddress;
