+++ src/cgw_connection_server.rs
@@ -1,7 +1,3 @@
-use crate::cgw_device::{
-    cgw_detect_device_changes, CGWDevice, CGWDeviceCapabilities, CGWDeviceState, CGWDeviceType,
-    OldNew,
-};
 use crate::cgw_nb_api_listener::{
     cgw_construct_cgw_alert_message, cgw_construct_cloud_header,
     cgw_construct_foreign_infra_connection_msg, cgw_construct_infra_capabilities_changed_msg,
@@ -15,15 +11,9 @@ use crate::cgw_nb_api_listener::{
     cgw_construct_unassigned_infra_leave_msg, cgw_get_timestamp_16_digits, CGWKafkaProducerTopic,
     ConsumerMetadata,
 };
-use crate::cgw_runtime::{cgw_get_runtime, CGWRuntimeType};
-use crate::cgw_tls::cgw_tls_get_cn_from_stream;
 use crate::cgw_ucentral_messages_queue_manager::{
     CGWUCentralMessagesQueueItem, CGWUCentralMessagesQueueState, CGW_MESSAGES_QUEUE,
     TIMEOUT_MANAGER_DURATION,
-};
-use crate::cgw_ucentral_parser::{
-    cgw_ucentral_parse_command_message, CGWUCentralCommand, CGWUCentralCommandType,
-    CGWUCentralConfigValidators,
 };
 use crate::cgw_ucentral_topology_map::CGWUCentralTopologyMap;
 use crate::AppArgs;
@@ -38,9 +28,21 @@ use crate::{
     },
     cgw_nb_api_listener::CGWNBApiClient,
     cgw_remote_discovery::CGWRemoteDiscovery,
+    cgw_runtime::{cgw_get_runtime, CGWRuntimeType},
 };
 
-use crate::cgw_errors::{Error, Result};
+use cgw_common::{
+    cgw_errors::{Error, Result},
+    cgw_tls::cgw_tls_get_cn_from_stream,
+    cgw_ucentral_parser::{
+        cgw_ucentral_parse_command_message, CGWUCentralCommand, CGWUCentralCommandType,
+        CGWUCentralConfigValidators,
+    },
+    cgw_device::{
+        cgw_detect_device_changes, CGWDevice, CGWDeviceCapabilities, CGWDeviceState, CGWDeviceType,
+        OldNew,
+    },
+};
 
 use std::str::FromStr;
 use std::{
@@ -3086,8 +3088,8 @@ mod tests {
     use crate::{
         cgw_errors::Result,
         cgw_ucentral_ap_parser::cgw_ucentral_ap_parse_message,
-        cgw_ucentral_parser::{CGWUCentralEvent, CGWUCentralEventType},
     };
+    use cgw_common::cgw_ucentral_parser::{CGWUCentralEvent, CGWUCentralEventType};
 
     fn get_connect_json_msg() -> &'static str {
         r#"
