+++ src/cgw_remote_discovery.rs
@@ -1,16 +1,19 @@
 use crate::{
-    cgw_app_args::CGWRedisArgs,
     cgw_db_accessor::{CGWDBAccessor, CGWDBInfra, CGWDBInfrastructureGroup},
-    cgw_device::{CGWDevice, CGWDeviceState, CGWDeviceType},
     cgw_devices_cache::CGWDevicesCache,
-    cgw_errors::{Error, Result},
     cgw_metrics::{
         CGWMetrics, CGWMetricsCounterOpType, CGWMetricsCounterType, CGWMetricsHealthComponent,
         CGWMetricsHealthComponentStatus,
     },
     cgw_remote_client::CGWRemoteClient,
-    cgw_tls::cgw_read_root_certs_dir,
     AppArgs,
+};
+
+use cgw_common::{
+    cgw_errors::{Error, Result},
+    cgw_app_args::CGWRedisArgs,
+    cgw_tls::cgw_read_root_certs_dir,
+    cgw_device::{CGWDevice, CGWDeviceState, CGWDeviceType},
 };
 
 use std::{
